+++ src/MR.AspNetCore.Jobs.SqlServer/SqlServerOptions.cs
@@ -5,5 +5,7 @@ namespace MR.AspNetCore.Jobs
 		public string ConnectionString { get; set; }
 
 		public string Schema { get; set; } = "Jobs";
+
+		public string MigrationsHistoryTableName { get; set; } = "__EFMigrationsHistory";
 	}
 }