+++ CHANGELOG.md
@@ -5,7 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
-_Nothing yet..._
+
+### Changed
+- SqlServer: Set migration history's table name and schema.
 
 ## [0.6.0] - 2017-04-22